client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["mcgillij <mcgillivray.jason@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12.0,<3.13"
discord-py = "^2.4.0"
redis = "^5.2.1"
pydub = "^0.25.1"
pynacl = "^1.5.0"
python-dotenv = "^1.0.1"
pyaudio = "^0.2.14"
wave = "^0.0.2"
aiohttp = "^3.11.11"
numpy = "^2.2.1"
discord-ext-voice-recv = "^0.4.2a145"
kokoro = "^0.9.2"
soundfile = "^0.13.1"
lmstudio = "^1.2.0"


[tool.poetry.group.dev.dependencies]
watchdog = "^6.0.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
